Disabled custom incoming call view + fix for normal incoming call view

+++ Linphone/Model/LinphoneManager.cs
@@ -642,10 +642,12 @@ namespace Linphone.Model
                 Logger.Msg("[LinphoneManager] Incoming received"); 
                 BaseModel.UIDispatcher.BeginInvoke(() =>
                 {
-                    BaseModel.CurrentPage.NavigationService.Navigate(new Uri("/Views/IncomingCall.xaml?sip=" + call.GetRemoteAddress().AsStringUriOnly(), UriKind.RelativeOrAbsolute));
-
-                    //Remove the current page from the back stack to avoid duplicating him after
-                    BaseModel.CurrentPage.NavigationService.RemoveBackEntry();
+                    if (Globals.Instance.CallController.CustomIncomingCallView)
+                    {
+                        BaseModel.CurrentPage.NavigationService.Navigate(new Uri("/Views/IncomingCall.xaml?sip=" + call.GetRemoteAddress().AsStringUriOnly(), UriKind.RelativeOrAbsolute));
+                        //Remove the current page from the back stack to avoid duplicating him after
+                        BaseModel.CurrentPage.NavigationService.RemoveBackEntry();
+                    }
 
                     LookupForContact(call);
                 });
